fix some errors
